ai-agent/requirements.txt
anthropic
openai>=1.40.0
playwright
pyautogui
pillow
python-dotenv
flask
apscheduler
chromadb
sentence-transformers
langdetect
pyperclip
hypothesis
pytest
watchdog
psutil
rumps
pywebview
